+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 * Improve the positioning of tooltips.
 * Only show tooltips if mouse is still.
+* `Slider` will now show the value display by default, unless turned off with `.show_value(false)`.
 
 
 ## 0.9.0 - 2021-02-07 - Light Mode and much more